More improvements to coin-node role
Define a skip-list, can be overridden by a run-list.
Follow up with Windows implementation.
Use local coin folder if found in $COIN_ROOT.

+++ minicoin/roles/coin-node/pre-provision.sh
@@ -3,7 +3,13 @@ cd $(dirname "$0")
 
 rm -rf coin
 
-git clone --single-branch --branch dev --depth 1 git://code.qt.io/qt/qt5.git
+if [[ ! -z $COIN_ROOT ]]
+then
+    cp -r $COIN_ROOT coin
+else
+    echo "COIN_ROOT not set, cloning from upstream"
+    git clone --single-branch --branch dev --depth 1 git://code.qt.io/qt/qt5.git
 
-mv qt5/coin coin
-rm -rf qt5
+    mv qt5/coin coin
+    rm -rf qt5
+fi
